creando los repositorios para autor y libro

+++ src/main/java/com/challenge/literatura/model/Autor.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class Autor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String nombre;
     private String fechaDeNacimiento;
 
@@ -22,7 +22,7 @@ public class Autor {
     }
 
     public Autor() {
-        
+
     }
 
     @Override
@@ -35,11 +35,11 @@ public class Autor {
                 '}';
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
